custom hook manager api
