+++ src/new-hooks/components/slider.js
@@ -1,34 +1,10 @@
 import React from 'react';
-import {
-  getSliderStyles,
-  getFrameStyles,
-  getImgTagStyles
-} from '../style-utils/slide-styles';
-
-const Slide = ({ child }) => {
-  return (
-    <div className="slider-slide">
-      {React.cloneElement(child, { style: getImgTagStyles() })}
-    </div>
-  );
-};
-
-export const slides = children => {
-  return children.map((child, index) => (
-    <Slide
-      child={child}
-      key={`${child.key}-${index}`}
-      frameStyles={getFrameStyles()}
-    />
-  ));
-};
+import { getSliderStyles } from '../style-utils/slide-styles';
 
 const Slider = props => {
-  const children = slides(props.children);
-
   return (
     <div className="slider" style={getSliderStyles(props.width, '50px')}>
-      {children[props.slideIndex]}
+      {props.slides}
     </div>
   );
 };
