+++ src/new-hooks/utils/manager-api.js
@@ -1,13 +1,13 @@
-import React, { useState, useEffect } from 'react';
 import _ from 'lodash';
-import { slides } from '../components/slider';
+import { slides } from '../components/slide';
+import React, { useState } from 'react';
 
-const SlideManager = props => {
-  const [slideIndex, setSlideIndex] = useState(2);
+const useSlideManager = props => {
+  const [slideIndex, setSlideIndex] = useState(props.slideIndex || 0);
   const slidesToShow = props.slidesToShow || 1;
   const children = slides(props.children);
   const slideCount = children.length;
-  console.log(children, slidesToShow);
+  console.log('slideIndex', slideIndex);
 
   const getCurrentSlides = () => {
     const currentSlides = _.slice(
@@ -15,14 +15,11 @@ const SlideManager = props => {
       slideIndex,
       slideIndex + slidesToShow
     );
-    console.log(currentSlides);
     return currentSlides;
   };
 
   const getPrevSlides = () => {
-    const prevSlide = slideIndex - 1;
     const prevSlides = _.slice(children, slideIndex - slidesToShow, slideIndex);
-    console.log(prevSlides);
     return prevSlides;
   };
 
@@ -32,24 +29,29 @@ const SlideManager = props => {
       slideIndex + 1,
       slideIndex + 1 + slidesToShow
     );
-    console.log(nextSlides);
     return nextSlides;
   };
 
   const goToSlide = ({ index }) => {
     return children[index];
   };
-  return [getCurrentSlides, getPrevSlides, getNextSlides, goToSlide];
+
+  const goToNext = () => {
+    setSlideIndex((slideIndex + 1) % slideCount);
+  };
+
+  const goToPrev = () => {
+    setSlideIndex((slideIndex - 1 + slideCount) % slideCount);
+  };
+
+  return {
+    currentSlides: getCurrentSlides(),
+    prevSlides: getPrevSlides(),
+    nextSlides: getNextSlides(),
+    goToNext,
+    goToPrev,
+    goToSlide
+  };
 };
 
-export default SlideManager;
-
-// export const getSlideIndex = () => {};
-
-// export const getCurrentSlide = () => {};
-
-// export const getPrevSlide = () => {};
-
-// export const getNextSlide = () => {};
-
-// export const getSlideCount = () => {};
+export default useSlideManager;
